requirements.txt
torch>=1.12.0
torchvision>=0.13.0
torchmetrics>=0.11.0
tqdm
Pillow
